Move RealmConfig.js
